Add registration to starter Add client to web-service Extract IAgentCommandApi
