+++ server/collector/pom.xml
@@ -55,6 +55,14 @@
       <groupId>com.google.guava</groupId>
       <artifactId>guava</artifactId>
     </dependency>
+
+    <!-- Collector is declared as a discoverable service,
+         so that its implementation can be invoked from other services to interacts with agents
+    -->
+    <dependency>
+      <groupId>org.bithon</groupId>
+      <artifactId>service-discovery-declaration</artifactId>
+    </dependency>
   </dependencies>
 
   <build>
