+++ server/service-discovery/client/src/main/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+org.bithon.server.discovery.client.DiscoveryClientAutoConfiguration
