+++ server/service-discovery/client/src/main/java/org/bithon/server/discovery/client/DiscoveryClientAutoConfiguration.java
@@ -1,0 +1,32 @@
+/*
+ *    Copyright 2020 bithon.org
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package org.bithon.server.discovery.client;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * @author Frank Chen
+ * @date 23/2/23 9:12 pm
+ */
+@Configuration(proxyBeanMethods = false)
+public class DiscoveryClientAutoConfiguration {
+    @Bean
+    ServiceBroadcastInvoker serviceBroadcastInvoker() {
+        return new ServiceBroadcastInvoker();
+    }
+}
