+++ server/service-discovery/registration/src/main/java/resources/META-INF/spring.factories
